requirements.txt
playwright==1.44.0
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
rapidfuzz==3.9.3
flask==3.0.3
geopy==2.4.1
Pillow==10.3.0
imagehash==4.3.1
python-dotenv==1.0.1
APScheduler==3.10.4
